+++ src/Util.h
@@ -7,6 +7,6 @@
 /**
  * Calcula la distancia entre dos puntos
  */
-int distance(const std::pair<int,int>& t1, const std::pair<int,int>& t2);
+double distance(const std::pair<int,int>& t1, const std::pair<int,int>& t2);
 
 #endif
